more typecasts
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/einit/trunk@2109 ce7afbd9-0f25-0410-bca2-9054d3de1fdb

+++ einit/src/modules/compatibility/compatibility-sysv-utmp.c
@@ -175,7 +175,7 @@ char updateutmp_f (enum utmp_action options, struct utmp *new_entry) {
        break;
 #ifdef DEBUG
       default:
-       notice (6, "bad UTMP entry: [%c%c%c%c] %i (%s), %s@%s: %i.%i\n", utmpentries[i].ut_id[0], utmpentries[i].ut_id[1], utmpentries[i].ut_id[2], utmpentries[i].ut_id[3], utmpentries[i].ut_type, utmpentries[i].ut_line, utmpentries[i].ut_user, utmpentries[i].ut_host, utmpentries[i].ut_tv.tv_sec, utmpentries[i].ut_tv.tv_usec);
+       notice (6, "bad UTMP entry: [%c%c%c%c] %i (%s), %s@%s: %i.%i\n", utmpentries[i].ut_id[0], utmpentries[i].ut_id[1], utmpentries[i].ut_id[2], utmpentries[i].ut_id[3], utmpentries[i].ut_type, utmpentries[i].ut_line, utmpentries[i].ut_user, utmpentries[i].ut_host, (int)utmpentries[i].ut_tv.tv_sec, (int)utmpentries[i].ut_tv.tv_usec);
        break;
 #endif
      }
